+++ src/main/java/com/milvus/vector_spring/project/dto/ProjectUpdateRequestDto.java
@@ -11,7 +11,7 @@ public class ProjectUpdateRequestDto {
     private String prompt;
     private String embedModel;
     private String basicModel;
-    private int dimensions;
+    private long dimensions;
 
     @NotNull
     @NotBlank
